feat: add shop and blog services
