+++ apps/cms/__tests__/blog.server.test.ts
@@ -1,5 +1,5 @@
 /** @jest-environment node */
-import { createPost, updatePost } from "../src/actions/blog.server";
+import { createPost, updatePost } from "../src/services/blog";
 
 jest.mock("../src/actions/common/auth", () => ({
   ensureAuthorized: jest.fn(),
@@ -17,34 +17,32 @@ jest.mock("@platform-core/src/shops", () => ({
   }),
 }));
 
+jest.mock("@platform-core/src/repositories/blog.server", () => ({
+  slugExists: jest.fn(),
+}));
+
 describe("blog post slug conflicts", () => {
-  const realFetch = global.fetch;
   afterEach(() => {
-    (global.fetch as any) = realFetch;
     jest.clearAllMocks();
   });
 
   it("returns error when slug already exists on create", async () => {
-    const fetchMock = jest.fn().mockResolvedValue({
-      json: async () => ({ result: [{ _id: "existing" }] }),
-    });
-    (global.fetch as any) = fetchMock;
+    const { slugExists } = require("@platform-core/src/repositories/blog.server");
+    (slugExists as jest.Mock).mockResolvedValue(true);
 
     const fd = new FormData();
     fd.set("title", "Title");
     fd.set("content", "[]");
     fd.set("slug", "dup");
 
-    const res = await createPost("shop", {} as any, fd);
+    const res = await createPost("shop", fd);
     expect(res).toEqual({ error: "Slug already exists" });
-    expect(fetchMock).toHaveBeenCalledTimes(1);
+    expect(slugExists).toHaveBeenCalledTimes(1);
   });
 
   it("returns error when slug already exists on update", async () => {
-    const fetchMock = jest.fn().mockResolvedValue({
-      json: async () => ({ result: [{ _id: "other" }] }),
-    });
-    (global.fetch as any) = fetchMock;
+    const { slugExists } = require("@platform-core/src/repositories/blog.server");
+    (slugExists as jest.Mock).mockResolvedValue(true);
 
     const fd = new FormData();
     fd.set("id", "current");
@@ -52,9 +50,9 @@ describe("blog post slug conflicts", () => {
     fd.set("content", "[]");
     fd.set("slug", "dup");
 
-    const res = await updatePost("shop", {} as any, fd);
+    const res = await updatePost("shop", fd);
     expect(res).toEqual({ error: "Slug already exists" });
-    expect(fetchMock).toHaveBeenCalledTimes(1);
+    expect(slugExists).toHaveBeenCalledTimes(1);
   });
 });
